Move httpie.manager.plugins to httpie.manager.tasks.plugins
